fix: split agent inference tests to satisfy backend-size limit
Mergeability follow-up after rebasing onto main: move inference-focused
agent executor tests into agent_inference_test.go so agent_test.go stays
under the 1000-line backend-size gate.
